+++ packages/api-fetch/src/middlewares/preloading.js
@@ -23,44 +23,54 @@ function createPreloadingMiddleware( preloadedData ) {
 				rawPath = pathFromQuery;
 			}
 		}
-		if ( typeof rawPath === 'string' ) {
-			const method = options.method || 'GET';
-			const path = normalizePath( rawPath );
+		if ( typeof rawPath !== 'string' ) {
+			return next( options );
+		}
 
-			if ( 'GET' === method && cache[ path ] ) {
-				const cacheData = cache[ path ];
+		const method = options.method || 'GET';
+		const path = normalizePath( rawPath );
 
-				// Unsetting the cache key ensures that the data is only used a single time
-				delete cache[ path ];
+		if ( 'GET' === method && cache[ path ] ) {
+			const cacheData = cache[ path ];
 
-				return Promise.resolve(
-					parse
-						? cacheData.body
-						: new window.Response(
-								JSON.stringify( cacheData.body ),
-								{
-									status: 200,
-									statusText: 'OK',
-									headers: cacheData.headers,
-								}
-						  )
-				);
-			} else if (
-				'OPTIONS' === method &&
-				cache[ method ] &&
-				cache[ method ][ path ]
-			) {
-				const cacheData = cache[ method ][ path ];
+			// Unsetting the cache key ensures that the data is only used a single time
+			delete cache[ path ];
 
-				// Unsetting the cache key ensures that the data is only used a single time
-				delete cache[ method ][ path ];
+			return prepareResponse( cacheData, !! parse );
+		} else if (
+			'OPTIONS' === method &&
+			cache[ method ] &&
+			cache[ method ][ path ]
+		) {
+			const cacheData = cache[ method ][ path ];
 
-				return Promise.resolve( parse ? cacheData.body : cacheData );
-			}
+			// Unsetting the cache key ensures that the data is only used a single time
+			delete cache[ method ][ path ];
+
+			return prepareResponse( cacheData, !! parse );
 		}
 
 		return next( options );
 	};
+}
+
+/**
+ * This is a helper function that sends a success response.
+ *
+ * @param {Record<string, any>} responseData
+ * @param {boolean}             parse
+ * @return {Promise<any>} Promise with the response.
+ */
+function prepareResponse( responseData, parse ) {
+	return Promise.resolve(
+		parse
+			? responseData.body
+			: new window.Response( JSON.stringify( responseData.body ), {
+					status: 200,
+					statusText: 'OK',
+					headers: responseData.headers,
+			  } )
+	);
 }
 
 export default createPreloadingMiddleware;