+++ packages/react-native-editor/CHANGELOG.md
@@ -10,6 +10,7 @@ For each user feature we should also add a importance categorization label  to i
 -->
 
 ## Unreleased
+-   [*] Image block: Replacing the media for an image set as featured prompts to update the featured image [#34666]
 
 -   [**] [Gallery block] Fix crash when adding images and selecting a gallery item [#38238]
 -   [***] Font size and line-height support for text-based blocks used in block-based themes [#38205]