+++ packages/react-native-editor/CHANGELOG.md
@@ -17,6 +17,12 @@ For each user feature we should also add a importance categorization label  to i
 
 ## 1.70.0
 
+## 1.70.1
+
+-   [**] [Gallery block] Fix crash when adding images and selecting a gallery item [#38238]
+
+## 1.70.0
+
 -   [**] Fix content justification attribute in Buttons block [#37887]
 -   [*] Hide help button from Unsupported Block Editor. [#37221]
 -   [*] Add contrast checker to text-based blocks [#34902]